aider: Left-justified "2048 Slide" title and right-justified "New Game" button in the header.
# Aider chat conversation:

USER: great, left justify "2048 Slide" and right justify "new game"

ASSISTANT: To left-justify the "2048 Slide" title and right-justify the "New Game" button, we need to modify the `2048slide.css` file to adjust the styling of the header. We will use flexbox properties to achieve the desired alignment.

Here are the *SEARCH/REPLACE* blocks for the `2048slide.css` file:

2048slide.css
```css
<<<<<<< SEARCH
header {
    display: flex;
    justify-content: center;
    align-items: center;
    margin-bottom: 20px;
    width: 100%;
    max-width: 700px; /* Assuming the game board width is 700px (100px * 7 tiles) */
    margin-left: auto;
    margin-right: auto;
}
=======
header {
    display: flex;
    justify-content: space-between;
    align-items: center;
    margin-bottom: 20px;
    width: 100%;
    max-width: 700px; /* Adjust this value to match the game board width */
    margin-left: auto;
    margin-right: auto;
}
>>>>>>> REPLACE
```

This change will make the header content spread out to the full width of the header container, with the title on the left and the button on the right. Make sure the `max-width` value matches the actual width of the game board. If the game board width is different from 700px, adjust the `max-width` value accordingly.